+++ static/genmon.js
@@ -1331,7 +1331,7 @@ function saveNotifications(){
              $('.progress-bar-fill').queue(function () {
                   $(this).css('width', '100%')
              });
-             setTimeout(function(){ vex.closeAll();}, 10000);
+             setTimeout(function(){ vex.closeAll();gotoLogin();}, 10000);
            }
         }
     })
@@ -2081,6 +2081,7 @@ function saveSettings(){
                 if ($('#fueltype').val() != $('#fueltype').attr('oldValue')) { myGenerator["fueltype"] = $('#fueltype').val(); }
                 if ($('#favicon').val() != $('#favicon').attr('oldValue')) { changeFavicon($('#favicon').val()); }
                 if (($('#enhancedexercise').prop('checked')  === true ? "true" : "false") != $('#enhancedexercise').attr('oldValue')) { myGenerator['EnhancedExerciseEnabled'] = ($('#enhancedexercise').prop('checked')  === true ? "true" : "false") }
+                gotoLogin();
              }, 10000);
            }
         }
@@ -2324,10 +2325,18 @@ function saveAddon(addon, addonTitle){
              });
              setTimeout(function(){
                 vex.closeAll();
+                gotoLogin();
              }, 10000);
+
            }
         }
     })
+}
+//*****************************************************************************
+function gotoLogin() {
+
+    var url = window.location.href.split("/")[0].split("?")[0];
+    window.location.href = url.concat("/logout");
 }
 //*****************************************************************************
 function saveAddonJSON(addon) {