+++ changelog.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Improvemnt in gensnmp.py to clean up on restarts (thanks @liltux)
 - Minor cleanup on genmqtt.py
 - Cleanup saving of settings web app redirect when using secure login (thanks @curtis1757)
+- Additional checks added for weather API city lookup
 
 ## V1.15.08 - 2020-10-27
 - Additional parameter validation for values read from conf file