+++ changelog.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## V1.15.10 - 2020-11-09
+- Added new tank sensor type for gentankdiy.py add on (thanks @curtis1757)
+- Misc code optimizations for maintenance purposes
+
 ## V1.15.09 - 2020-11-05
 - Rounded value written to power log to 3 decimal places (Evolution)
 - Corrected issue with email errors relating to RFC2821 (see https://tools.ietf.org/html/rfc2821#section-3.3)